BAU: Change subclasses of TrustStoreConfiguration to use common name
Changed both subclasses to use "store" as the common name and aliased
previous name for the file backed config (for backwards compat reasons)
As no apps are currently using the encoded trust stores, this can be freely
changed. The purpose of this change was so that anyone overriding config
only needs to change the "type" field in YAML to change the format without
also having to change to a new key.

Authors: @andy-paine