BAU: Cut dependency of saml-metadata-bindings on saml-test-utils.
We only include saml-test-utils to gain access to a small number of builders.
For these few of builders, we also end up including a large amount of
code and libraries that we don't need. This makes updating any of the
libraries much harder than it needs to be.

As a result, we have taken the descision to duplicate some builders in this
project, so that we are able to reduce our dependencies.

(note that the tests for saml-metadata-bindings had a transitive dependency
on ida-saml-extensions via saml-metadata-bindings -> saml-metadata-bindings-test
-> saml-test-utils -> saml-domain-objects -> ida-saml-extensions/
This dependency is now declared explicitly)

Authors: @rachelthecodesmith, @michaelwalker