BAU: Cut dependency on saml-domain-objects
saml-test-utils previous contained several builders for classes defined
in saml-domain-objects.
These builders have now been pushed down into saml-domain-objects, and
so we are largely able to cleanly cut the dependency.

The one downside of this is the need to add a copy of the (already deprecated)
OpenSamlXmlObjectFactory from saml-domain-objects.
saml-test-utils now contains a (reduced and still deprecated) version of
the original class.

Solo: @michaelwalker