EID-428: Allow naming of MetadataHealthCheck
You can use the `/healthcheck` endpoint of our Dropwizard apps that use
saml-metadata-bindings to check on whether metadata was successfully
consumed. However, by default the name of the healthcheck was set to
"metadata" meaning if you're consuming metadata from multiple sources
in your application, only one healthcheck will be returned from the
endpoint, making it impossible to determine which
MetadataResolver might be failing.

Author: @vixus0