refactor: Use OpenSAML helpers for deserializing input
We were using our own helpers for parsing XML from strings and Saml from
XML. This was largely a duplicate of similar code that OpenSAML
provides and in some cases it was doing things less efficiently or with
fewer checks.

The key change has been that `SamlObjectParser` will now use
 `XMLObjectSupport` and `ParserPool` to convert from Strings to SAML
 Objects. `SamlObjectParser` will still convert from `Element` to SAML
objects in the same way, but it is unclear whether clients really need
to do this.