EID-428: Allow changing EncrypterFactory options
To use EncrypterFactory for non-Verify purposes such as eIDAS, we need
to be able to change the encryption algorithms and key placement used. The
`withAes256Encryption()` method was only used in one place (Stub IDP)
for eIDAS reasons so it should be changed there to use the more general
method.

Author: @vixus0