firebreak: API for metadata SignatureTrustEngine
Wherever we validate signatures using opensaml libraries we need to
create a SignatureTrustEngine object that does validation and credential
fetching. When using creds from Metadata we have so far duplicated the
building of these objects across our codebase. This change introduces
a single way of doing this that can be used everywhere, which should
make our code easier to understand and support.