firebreak: add SignatureTrustEngine to eidas metadata repository
When validating xml signatures using OpenSaml we need
a SignatureTrustEngine object. We have so far produced many different
ways of doing this which adds to a complicated and hard to understand
codebase. This change makes the EidasMetadataResolverRepository use the
new MetadataSignatureTrustEngineFactory object that should be the way we
validate signatures where we need to resolve creds from saml metadata.