firebreak: API for MetadataCredentialResolver
When encrypting SAML messages we need an Encryption credential. This
change introduces an API for providing a metadata CredentialResolver
that can be used across our codebase to provide a single way of
resolving encryption credentials from metadata so that we don't end up
with many different ways of doing it.